Split up routes and start on initiative improvements
